+++ packages/boxel-ui/addon/src/components/resizable-panel-group/utils/adjust-layout-by-delta.ts
@@ -1,6 +1,6 @@
-import { fuzzyLayoutsEqual } from './fuzzyLayoutsEqual.ts';
-import { fuzzyNumbersEqual } from './fuzzyNumbers.ts';
-import { resizePanel } from './resizePanel.ts';
+import { fuzzyLayoutsEqual } from './fuzzy-layouts-equal.ts';
+import { fuzzyNumbersEqual } from './fuzzy-numbers.ts';
+import { resizePanel } from './resize-panel.ts';
 import type { ResizablePanelConstraints } from './types.ts';
 
 export function adjustLayoutByDelta({
