+++ packages/boxel-ui/addon/src/components/resizable-panel-group/utils/calculate-unsafe-default-layout.ts
