+++ packages/boxel-ui/addon/src/components/resizable-panel-group/utils/compare-layouts.ts
