+++ packages/boxel-ui/addon/src/components/resizable-panel-group/utils/dom/get-panel-elements-for-group.ts
