+++ packages/boxel-ui/addon/src/components/resizable-panel-group/utils/dom/get-resize-handle-element-index.ts
@@ -1,4 +1,4 @@
-import { getResizeHandleElementsForGroup } from './getResizeHandleElementsForGroup.ts';
+import { getResizeHandleElementsForGroup } from './get-resize-handle-elements-for-group.ts';
 
 export function getResizeHandleElementIndex(
   groupId: string,
