+++ packages/boxel-ui/addon/src/components/resizable-panel-group/utils/dom/get-resize-handle-element.ts
