+++ packages/boxel-ui/addon/src/components/resizable-panel-group/utils/get-resize-event-coordinates.ts
